streamflow/services/auth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.0.0
pika==1.3.2